backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langsmith>=0.3.0
